lockfile

+++ src/lockfile.iced
@@ -35,15 +35,16 @@ class Lockfile
 
   #--------------------
 
-  constructor : ({@filename, wait_limit, poke_timeout, poke_interval, retry_interval, mode, @log}) ->
-    @retry_interval = retry_interval or 100                 # Retry every interval ms
-    @poke_interval  = poke_interval or 100                  # poke the file every 100ms
-    @poke_timeout   = poke_timeout or (@poke_interval * 20) # after 20 missed pokes, declare failure
-    @wait_limit     = 10*1000                               # give up the wait after 10s
-    @mode           = mode or 0o644                         # prefered file mode 
-    @id             = get_id()
-    @_locked        = false
-    @_maintain_cb   = null
+  constructor : ({@filename, wait_limit, poke_timeout, reclaim_timeout, poke_interval, retry_interval, mode, @log}) ->
+    @retry_interval  = retry_interval or 100                  # Retry every interval ms
+    @poke_interval   = poke_interval or 100                   # poke the file every 100ms
+    @poke_timeout    = poke_timeout or (@poke_interval * 20)  # after 20 missed pokes, declare failure
+    @reclaim_timeout = reclaim_timeout or (@poke_interval*5)  # After 5 timeslots, we can take it
+    @wait_limit      = 10*1000                                # give up the wait after 10s
+    @mode            = mode or 0o644                          # prefered file mode 
+    @id              = get_id()
+    @_locked         = false
+    @_maintain_cb    = null
 
   #--------------------
 
@@ -61,12 +62,13 @@ class Lockfile
 
   _acquire_1 : (cb) ->
     res = false
+    unlinked = false
     await fs.open @filename, 'wx', @mode, defer err, @fd
     if not err? then res = true
     else
-      await @_acquire_1_fallback defer err
+      await @_acquire_1_fallback defer err, unlinked
       @warn err.message if err?
-    cb res
+    cb res, unlinked
 
   #--------------------
 
@@ -76,21 +78,31 @@ class Lockfile
 
   #--------------------
 
+  # If we failed to acquire the lock, we now need to ask if the lock is stale
+  # because its previous holder died while holding it.  We do this by:
+  #   1. Checking the timestamp in the lock.
+  #   2. If the timestamp is too far in the past, then we try to overwrite the lock with
+  #      our process's info. Note that this operation isn't guaranteed to be atomic,
+  #      and we'll have no way of knowing if we won or not.
+  #   3. If our overwrite stands for a short timeout, then we can conclude that we won.
+  #      Now, we're allowed to delete the file.
   _acquire_1_fallback : (cb) ->
+    unlinked = false
     esc = make_esc cb, "_acquire_1_fallback"
     err = null
     await fs.open @filename, 'r', esc defer rfd
     await read_all rfd, esc defer buf
     await a_json_parse buf, esc defer jso
+    now = Date.now()
     if not Array.isArray(jso) or jso.length < 2
       err = new Error "Bad lock file; expected an array with 2 values"
-    else if (Date.now() - jso[0]) < @poke_timeout then # noop, still locked?
-    else if jso[1] is @id
+    else if (jso[1] is @id) and (now - jso[0] > @reclaim_timeout)
       await fs.unlink @filename, esc defer()
-    else
+      unlinked = true
+    else if (jso[1] isnt @id) and (now - jso[0] > @poke_timeout)
       obj = @_lock_dat() 
       await fs.writeFile @filename, obj, { encoding : 'utf8', @mode}, esc defer()
-    cb null
+    cb null, unlinked
 
   #--------------------
 
@@ -98,10 +110,10 @@ class Lockfile
     acquired = false
     start = Date.now()
     loop
-      await @_acquire_1 defer acquired
+      await @_acquire_1 defer acquired, unlinked
       break if acquired
       break if @wait_limit and (Date.now() - start) > @wait_limit
-      await setTimeout defer(), @retry_interval
+      await setTimeout defer(), @retry_interval unless unlinked
     if acquired
       @maintain_lock_loop()
     cb acquired
@@ -143,14 +155,3 @@ class Lockfile
 exports.Lockfile = Lockfile
 
 #==========================================================================
-
-lock = new Lockfile { filename : "/tmp/shit" }
-console.log "acquire...."
-await lock.acquire defer t
-console.log t
-await setTimeout defer(), 1000
-lock.release()
-await setTimeout defer(), 100
-process.exit 1
-
-
